Migrated auth/create-user.js to typescript.
